+++ library/src/iusReceiveChannelMapDict.c
@@ -121,7 +121,7 @@ size_t iusReceiveChannelMapDictGetSize
 	iurcmd_t dict
 )
 {
-	IUS_ERR_CHECK_NULL_N_RETURN(dict, -1);
+	IUS_ERR_CHECK_NULL_N_RETURN(dict, (size_t) -1);
 	return hashmap_size(&dict->map);
 }
